`linera-rpc`: feature-gate server functionality
