+++ linera-rpc/Cargo.toml
@@ -29,7 +29,16 @@ metrics = [
     "linera-views/metrics",
 ]
 
-simple-network = []
+server = ["tonic-health"]
+simple-network = ["tokio-util/net"]
+
+web = [
+    "linera-base/web",
+    "linera-chain/web",
+    "linera-core/web",
+    "linera-execution/web",
+    "linera-views/web",
+]
 
 [dependencies]
 anyhow.workspace = true
@@ -56,6 +65,8 @@ rand.workspace = true
 serde.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
+tokio-util = { workspace = true, optional = true, features = ["codec"] }
+tonic-health = { workspace = true, optional = true }
 tower.workspace = true
 tracing.workspace = true
 
@@ -68,14 +79,10 @@ similar-asserts.workspace = true
 test-strategy.workspace = true
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-tokio.workspace = true
-tokio-util = { workspace = true, features = ["codec", "net"] }
-tonic = { workspace = true, features = ["codegen", "prost", "tls"] }
-tonic-health.workspace = true
-tower.workspace = true
+tonic = { workspace = true, features = ["prost", "codegen", "transport"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-tonic = { workspace = true, features = ["codegen", "prost"] }
+tonic = { workspace = true, features = ["prost", "codegen"] }
 
 [build-dependencies]
 cfg_aliases.workspace = true
