+++ linera-rpc/build.rs
@@ -10,6 +10,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     cfg_aliases::cfg_aliases! {
         web: { all(target_arch = "wasm32", target_os = "unknown") },
         with_metrics: { all(not(web), feature = "metrics") },
+        with_server: { all(not(web), feature = "server") },
         with_simple_network: { all(not(web), feature = "simple-network") },
     };
