+++ linera-rpc/src/grpc/client.rs
@@ -1,0 +1,306 @@
+// Copyright (c) Zefchain Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use super::{
+    api::{
+        chain_info_result::Inner, validator_node_client::ValidatorNodeClient, SubscriptionRequest,
+    },
+    Error, ProtoConversionError, MAX_MESSAGE_SIZE,
+};
+use crate::{
+    config::ValidatorPublicNetworkConfig,
+    mass::{MassClient, MassClientError},
+    node_provider::NodeOptions,
+    HandleCertificateRequest, HandleLiteCertificateRequest, RpcMessage,
+};
+use async_trait::async_trait;
+use futures::{future, stream, StreamExt};
+use linera_base::identifiers::ChainId;
+use linera_chain::data_types;
+use linera_core::{
+    node::{CrossChainMessageDelivery, NodeError, NotificationStream, ValidatorNode},
+    worker::Notification,
+};
+
+use linera_version::VersionInfo;
+
+use std::{iter, time::Duration};
+
+use tonic::{Code, Request, Status};
+
+use tracing::{debug, info, instrument, warn};
+
+#[derive(Clone)]
+pub struct Client {
+    address: String,
+    client: ValidatorNodeClient<tonic::transport::Channel>,
+    notification_retry_delay: Duration,
+    notification_retries: u32,
+}
+
+impl Client {
+    pub fn new(network: ValidatorPublicNetworkConfig, options: NodeOptions) -> Result<Self, Error> {
+        let address = network.http_address();
+        let channel = tonic::transport::Channel::from_shared(address.clone())?
+            .connect_timeout(options.send_timeout)
+            .timeout(options.recv_timeout)
+            .connect_lazy();
+        let client = ValidatorNodeClient::new(channel)
+            .max_encoding_message_size(MAX_MESSAGE_SIZE)
+            .max_decoding_message_size(MAX_MESSAGE_SIZE);
+
+        Ok(Self {
+            address,
+            client,
+            notification_retry_delay: options.notification_retry_delay,
+            notification_retries: options.notification_retries,
+        })
+    }
+
+    /// Returns whether this gRPC status means the server stream should be reconnected to, or not.
+    /// Logs a warning on unexpected status codes.
+    fn is_retryable(status: &Status) -> bool {
+        match status.code() {
+            Code::DeadlineExceeded | Code::Aborted | Code::Unavailable | Code::Unknown => {
+                info!("Notification stream interrupted: {}; retrying", status);
+                true
+            }
+            Code::Ok
+            | Code::Cancelled
+            | Code::NotFound
+            | Code::AlreadyExists
+            | Code::ResourceExhausted => {
+                warn!("Unexpected gRPC status: {}; retrying", status);
+                true
+            }
+            Code::InvalidArgument
+            | Code::PermissionDenied
+            | Code::FailedPrecondition
+            | Code::OutOfRange
+            | Code::Unimplemented
+            | Code::Internal
+            | Code::DataLoss
+            | Code::Unauthenticated => {
+                warn!("Unexpected gRPC status: {}", status);
+                false
+            }
+        }
+    }
+}
+
+macro_rules! client_delegate {
+    ($self:ident, $handler:ident, $req:ident) => {{
+        debug!(request = ?$req, "sending gRPC request");
+        let request_inner = $req.try_into().map_err(|_| NodeError::GrpcError {
+            error: "could not convert request to proto".to_string(),
+        })?;
+        let request = Request::new(request_inner);
+        match $self
+            .client
+            .$handler(request)
+            .await
+            .map_err(|s| NodeError::GrpcError {
+                error: format!(
+                    "remote request [{}] failed with status: {:?}",
+                    stringify!($handler),
+                    s
+                ),
+            })?
+            .into_inner()
+            .inner
+            .ok_or(NodeError::GrpcError {
+                error: "missing body from response".to_string(),
+            })? {
+            Inner::ChainInfoResponse(response) => {
+                Ok(response.try_into().map_err(|err| NodeError::GrpcError {
+                    error: format!("failed to marshal response: {}", err),
+                })?)
+            }
+            Inner::Error(error) => {
+                Err(bincode::deserialize(&error).map_err(|err| NodeError::GrpcError {
+                    error: format!("failed to marshal error message: {}", err),
+                })?)
+            }
+        }
+    }};
+}
+
+#[async_trait]
+impl ValidatorNode for Client {
+    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
+    async fn handle_block_proposal(
+        &mut self,
+        proposal: data_types::BlockProposal,
+    ) -> Result<linera_core::data_types::ChainInfoResponse, NodeError> {
+        client_delegate!(self, handle_block_proposal, proposal)
+    }
+
+    #[instrument(target = "grpc_client", skip_all, fields(address = self.address))]
+    async fn handle_lite_certificate(
+        &mut self,
+        certificate: data_types::LiteCertificate<'_>,
+        delivery: CrossChainMessageDelivery,
+    ) -> Result<linera_core::data_types::ChainInfoResponse, NodeError> {
+        let wait_for_outgoing_messages = delivery.wait_for_outgoing_messages();
+        let request = HandleLiteCertificateRequest {
+            certificate,
+            wait_for_outgoing_messages,
+        };
+        client_delegate!(self, handle_lite_certificate, request)
+    }
+
+    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
+    async fn handle_certificate(
+        &mut self,
+        certificate: data_types::Certificate,
+        blobs: Vec<data_types::HashedValue>,
+        delivery: CrossChainMessageDelivery,
+    ) -> Result<linera_core::data_types::ChainInfoResponse, NodeError> {
+        let wait_for_outgoing_messages = delivery.wait_for_outgoing_messages();
+        let request = HandleCertificateRequest {
+            certificate,
+            blobs,
+            wait_for_outgoing_messages,
+        };
+        client_delegate!(self, handle_certificate, request)
+    }
+
+    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
+    async fn handle_chain_info_query(
+        &mut self,
+        query: linera_core::data_types::ChainInfoQuery,
+    ) -> Result<linera_core::data_types::ChainInfoResponse, NodeError> {
+        client_delegate!(self, handle_chain_info_query, query)
+    }
+
+    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
+    async fn subscribe(&mut self, chains: Vec<ChainId>) -> Result<NotificationStream, NodeError> {
+        let notification_retry_delay = self.notification_retry_delay;
+        let notification_retries = self.notification_retries;
+        let mut retry_count = 0;
+        let subscription_request = SubscriptionRequest {
+            chain_ids: chains.into_iter().map(|chain| chain.into()).collect(),
+        };
+        let mut client = self.client.clone();
+
+        // Make the first connection attempt before returning from this method.
+        let mut stream = Some(
+            client
+                .subscribe(subscription_request.clone())
+                .await
+                .map_err(|status| NodeError::SubscriptionFailed {
+                    status: status.to_string(),
+                })?
+                .into_inner(),
+        );
+
+        // A stream of `Result<grpc::Notification, tonic::Status>` that keeps calling
+        // `client.subscribe(request)` endlessly and without delay.
+        let endlessly_retrying_notification_stream = stream::unfold((), move |()| {
+            let mut client = client.clone();
+            let subscription_request = subscription_request.clone();
+            let mut stream = stream.take();
+            async move {
+                let stream = if let Some(stream) = stream.take() {
+                    future::Either::Right(stream)
+                } else {
+                    match client.subscribe(subscription_request.clone()).await {
+                        Err(err) => future::Either::Left(stream::iter(iter::once(Err(err)))),
+                        Ok(response) => future::Either::Right(response.into_inner()),
+                    }
+                };
+                Some((stream, ()))
+            }
+        })
+        .flatten();
+
+        // The stream of `Notification`s that inserts increasing delays after retriable errors, and
+        // terminates after unexpected or fatal errors.
+        let notification_stream = endlessly_retrying_notification_stream
+            .map(|result| {
+                Notification::try_from(result?).map_err(|err| {
+                    let message = format!("Could not deserialize notification: {}", err);
+                    tonic::Status::new(Code::Internal, message)
+                })
+            })
+            .take_while(move |result| {
+                let Err(status) = result else {
+                    retry_count = 0;
+                    return future::Either::Left(future::ready(true));
+                };
+                if !Self::is_retryable(status) || retry_count >= notification_retries {
+                    return future::Either::Left(future::ready(false));
+                }
+                let delay = notification_retry_delay.saturating_mul(retry_count);
+                retry_count += 1;
+                future::Either::Right(async move {
+                    tokio::time::sleep(delay).await;
+                    true
+                })
+            })
+            .filter_map(|result| future::ready(result.ok()));
+
+        Ok(Box::pin(notification_stream))
+    }
+
+    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
+    async fn get_version_info(&mut self) -> Result<VersionInfo, NodeError> {
+        Ok(self.client.get_version_info(()).await?.into_inner().into())
+    }
+}
+
+#[async_trait]
+impl MassClient for Client {
+    #[allow(clippy::blocks_in_conditions)]
+    #[tracing::instrument(skip_all, err)]
+    async fn send(
+        &mut self,
+        requests: Vec<RpcMessage>,
+        max_in_flight: usize,
+    ) -> Result<Vec<RpcMessage>, MassClientError> {
+        let client = &mut self.client;
+        let responses = stream::iter(requests)
+            .map(|request| {
+                let mut client = client.clone();
+                async move {
+                    let response = match request {
+                        RpcMessage::BlockProposal(proposal) => {
+                            let request = Request::new((*proposal).try_into()?);
+                            client.handle_block_proposal(request).await?
+                        }
+                        RpcMessage::Certificate(request) => {
+                            let request = Request::new((*request).try_into()?);
+                            client.handle_certificate(request).await?
+                        }
+                        msg => panic!("attempted to send msg: {:?}", msg),
+                    };
+                    match response
+                        .into_inner()
+                        .inner
+                        .ok_or(ProtoConversionError::MissingField)?
+                    {
+                        Inner::ChainInfoResponse(chain_info_response) => {
+                            Ok(Some(RpcMessage::ChainInfoResponse(Box::new(
+                                chain_info_response.try_into()?,
+                            ))))
+                        }
+                        Inner::Error(error) => {
+                            let error = bincode::deserialize::<NodeError>(&error)
+                                .map_err(ProtoConversionError::BincodeError)?;
+                            tracing::error!(?error, "received error response");
+                            Ok(None)
+                        }
+                    }
+                }
+            })
+            .buffer_unordered(max_in_flight)
+            .filter_map(
+                |result: Result<Option<_>, MassClientError>| async move { result.transpose() },
+            )
+            .collect::<Vec<_>>()
+            .await
+            .into_iter()
+            .collect::<Result<Vec<_>, _>>()?;
+        Ok(responses)
+    }
+}
