+++ linera-rpc/src/grpc/mod.rs
@@ -1,14 +1,40 @@
-mod network;
-pub mod pool;
-mod node_provider;
+mod client;
 mod conversions;
+mod node_provider;
+pub mod pool;
 
-pub use network::*;
-pub use node_provider::*;
+#[cfg(with_server)]
+mod server;
+
+pub use client::*;
 pub use conversions::*;
+pub use node_provider::*;
+#[cfg(with_server)]
+pub use server::*;
 
 #[allow(clippy::derive_partial_eq_without_eq)]
 // https://github.com/hyperium/tonic/issues/1056
 pub mod api {
     tonic::include_proto!("rpc.v1");
 }
+
+#[derive(thiserror::Error, Debug)]
+pub enum Error {
+    #[error("failed to connect to address: {0}")]
+    ConnectionFailed(#[from] tonic::transport::Error),
+
+    #[error("failed to communicate cross-chain queries: {0}")]
+    CrossChain(#[from] tonic::Status),
+
+    #[error("failed to execute task to completion: {0}")]
+    Join(#[from] tokio::task::JoinError),
+
+    #[error("failed to parse socket address: {0}")]
+    SocketAddr(#[from] std::net::AddrParseError),
+
+    #[error(transparent)]
+    InvalidUri(#[from] http::uri::InvalidUri),
+}
+
+const MEBIBYTE: usize = 1024 * 1024;
+pub const MAX_MESSAGE_SIZE: usize = 8 * MEBIBYTE;
