+++ linera-rpc/src/grpc/server.rs
@@ -1,71 +1,49 @@
-// Copyright (c) Zefchain Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
 // `tracing::instrument` is not compatible with this nightly Clippy lint
 #![allow(unknown_lints)]
 #![allow(clippy::blocks_in_conditions)]
 
-#[allow(clippy::derive_partial_eq_without_eq)]
-// https://github.com/hyperium/tonic/issues/1056
-pub mod grpc {
-    tonic::include_proto!("rpc.v1");
-}
+use super::pool::ConnectionPool;
 
 use crate::{
-    config::{
-        CrossChainConfig, NotificationConfig, ShardId, ValidatorInternalNetworkConfig,
-        ValidatorPublicNetworkConfig,
-    },
-    conversions::ProtoConversionError,
-    grpc_pool::ConnectionPool,
-    mass::{MassClient, MassClientError},
-    node_provider::NodeOptions,
-    rpc::{HandleCertificateRequest, HandleLiteCertificateRequest},
-    RpcMessage,
+    config::{CrossChainConfig, NotificationConfig, ShardId, ValidatorInternalNetworkConfig},
+    HandleCertificateRequest, HandleLiteCertificateRequest,
 };
-use async_trait::async_trait;
+
+use super::{
+    api::{
+        self,
+        notifier_service_client::NotifierServiceClient,
+        validator_worker_client::ValidatorWorkerClient,
+        validator_worker_server::{ValidatorWorker as ValidatorWorkerRpc, ValidatorWorkerServer},
+        BlockProposal, Certificate, ChainInfoQuery, ChainInfoResult, CrossChainRequest,
+        LiteCertificate,
+    },
+    Error, MAX_MESSAGE_SIZE,
+};
 use futures::{
     channel::{mpsc, mpsc::Receiver, oneshot::Sender},
-    future,
     future::BoxFuture,
-    stream, FutureExt, StreamExt,
-};
-use grpc::{
-    chain_info_result::Inner,
-    notifier_service_client::NotifierServiceClient,
-    validator_node_client::ValidatorNodeClient,
-    validator_worker_client::ValidatorWorkerClient,
-    validator_worker_server::{ValidatorWorker as ValidatorWorkerRpc, ValidatorWorkerServer},
-    BlockProposal, Certificate, ChainInfoQuery, ChainInfoResult, CrossChainRequest,
-    LiteCertificate, SubscriptionRequest,
+    FutureExt, StreamExt,
 };
 use linera_base::identifiers::ChainId;
-use linera_chain::data_types;
+
 use linera_core::{
-    node::{CrossChainMessageDelivery, NodeError, NotificationStream, ValidatorNode},
+    node::NodeError,
     worker::{NetworkActions, Notification, ValidatorWorker, WorkerError, WorkerState},
 };
 use linera_storage::Storage;
-use linera_version::VersionInfo;
+
 use linera_views::views::ViewError;
 use rand::Rng;
 use std::{
-    fmt::Debug,
-    iter,
-    net::{AddrParseError, SocketAddr},
+    net::SocketAddr,
     str::FromStr,
     task::{Context, Poll},
     time::{Duration, Instant},
 };
-use thiserror::Error;
-use tokio::{
-    sync::oneshot,
-    task::{JoinError, JoinHandle},
-};
-use tonic::{
-    transport::{Body, Channel, Server},
-    Code, Request, Response, Status,
-};
+
+use tokio::{sync::oneshot, task::JoinHandle};
+use tonic::{Request, Response, Status};
 use tower::{builder::ServiceBuilder, Layer, Service};
 use tracing::{debug, error, info, instrument, warn};
 
@@ -74,9 +52,6 @@ use {
     linera_base::{prometheus_util, sync::Lazy},
     prometheus::{HistogramVec, IntCounterVec},
 };
-
-const MEBIBYTE: usize = 1024 * 1024;
-pub const MAX_MESSAGE_SIZE: usize = 8 * MEBIBYTE;
 
 type CrossChainSender = mpsc::Sender<(linera_core::data_types::CrossChainRequest, ShardId)>;
 type NotificationSender = mpsc::Sender<Notification>;
@@ -130,7 +105,7 @@ static SERVER_REQUEST_LATENCY_PER_REQUEST_TYPE: Lazy<HistogramVec> = Lazy::new(|
 });
 
 #[derive(Clone)]
-pub struct GrpcServer<S> {
+pub struct Server<S> {
     state: WorkerState<S>,
     shard_id: ShardId,
     network: ValidatorInternalNetworkConfig,
@@ -138,36 +113,15 @@ pub struct GrpcServer<S> {
     notification_sender: NotificationSender,
 }
 
-pub struct GrpcServerHandle {
+pub struct ServerHandle {
     _complete: Sender<()>,
     handle: JoinHandle<Result<(), tonic::transport::Error>>,
 }
 
-impl GrpcServerHandle {
-    pub async fn join(self) -> Result<(), GrpcError> {
+impl ServerHandle {
+    pub async fn join(self) -> Result<(), Error> {
         Ok(self.handle.await??)
     }
-}
-
-#[derive(Error, Debug)]
-pub enum GrpcError {
-    #[error("failed to connect to address: {0}")]
-    ConnectionFailed(#[from] tonic::transport::Error),
-
-    #[error("failed to convert to proto: {0}")]
-    ProtoConversion(#[from] ProtoConversionError),
-
-    #[error("failed to communicate cross-chain queries: {0}")]
-    CrossChain(#[from] Status),
-
-    #[error("failed to execute task to completion: {0}")]
-    Join(#[from] JoinError),
-
-    #[error("failed to parse socket address: {0}")]
-    SocketAddr(#[from] AddrParseError),
-
-    #[error(transparent)]
-    InvalidUri(#[from] http::uri::InvalidUri),
 }
 
 #[derive(Clone)]
@@ -186,10 +140,11 @@ impl<S> Layer<S> for PrometheusMetricsMiddlewareLayer {
     }
 }
 
-impl<S> Service<tonic::codegen::http::Request<Body>> for PrometheusMetricsMiddlewareService<S>
+impl<S> Service<tonic::codegen::http::Request<tonic::transport::Body>>
+    for PrometheusMetricsMiddlewareService<S>
 where
     S::Future: Send + 'static,
-    S: Service<tonic::codegen::http::Request<Body>> + std::marker::Send,
+    S: Service<tonic::codegen::http::Request<tonic::transport::Body>> + std::marker::Send,
 {
     type Response = S::Response;
     type Error = S::Error;
@@ -199,7 +154,7 @@ where
         self.service.poll_ready(cx)
     }
 
-    fn call(&mut self, request: http::Request<Body>) -> Self::Future {
+    fn call(&mut self, request: http::Request<tonic::transport::Body>) -> Self::Future {
         #[cfg(with_metrics)]
         let start = Instant::now();
         let future = self.service.call(request);
@@ -218,7 +173,7 @@ where
     }
 }
 
-impl<S> GrpcServer<S>
+impl<S> Server<S>
 where
     S: Storage + Clone + Send + Sync + 'static,
     ViewError: From<S::ContextError>,
@@ -232,7 +187,7 @@ where
         internal_network: ValidatorInternalNetworkConfig,
         cross_chain_config: CrossChainConfig,
         notification_config: NotificationConfig,
-    ) -> Result<GrpcServerHandle, GrpcError> {
+    ) -> Result<ServerHandle, Error> {
         info!(
             "spawning gRPC server on {}:{} for shard {}",
             host, port, shard_id
@@ -283,7 +238,7 @@ where
             .set_serving::<ValidatorWorkerServer<Self>>()
             .await;
 
-        let grpc_server = GrpcServer {
+        let grpc_server = Server {
             state,
             shard_id,
             network: internal_network,
@@ -296,7 +251,7 @@ where
             .max_decoding_message_size(MAX_MESSAGE_SIZE);
 
         let handle = tokio::spawn(
-            Server::builder()
+            tonic::transport::Server::builder()
                 .layer(
                     ServiceBuilder::new()
                         .layer(PrometheusMetricsMiddlewareLayer)
@@ -307,7 +262,7 @@ where
                 .serve_with_shutdown(server_address, receiver.map(|_| ())),
         );
 
-        Ok(GrpcServerHandle {
+        Ok(ServerHandle {
             _complete: complete,
             handle,
         })
@@ -321,7 +276,7 @@ where
         proxy_address: String,
         mut receiver: Receiver<Notification>,
     ) {
-        let channel = Channel::from_shared(proxy_address.clone())
+        let channel = tonic::transport::Channel::from_shared(proxy_address.clone())
             .expect("Proxy URI should be valid")
             .connect_lazy();
         let mut client = NotifierServiceClient::new(channel)
@@ -329,14 +284,14 @@ where
             .max_decoding_message_size(MAX_MESSAGE_SIZE);
 
         while let Some(notification) = receiver.next().await {
-            let notification: grpc::Notification = match notification.clone().try_into() {
+            let notification: api::Notification = match notification.clone().try_into() {
                 Ok(notification) => notification,
                 Err(error) => {
                     warn!(%error, nickname, "could not deserialize notification");
                     continue;
                 }
             };
-            let request = Request::new(notification.clone());
+            let request = tonic::Request::new(notification.clone());
             if let Err(error) = client.notify(request).await {
                 error!(
                     %error,
@@ -470,7 +425,7 @@ where
 }
 
 #[tonic::async_trait]
-impl<S> ValidatorWorkerRpc for GrpcServer<S>
+impl<S> ValidatorWorkerRpc for Server<S>
 where
     S: Storage + Clone + Send + Sync + 'static,
     ViewError: From<S::ContextError>,
@@ -645,283 +600,6 @@ where
     }
 }
 
-#[derive(Clone)]
-pub struct GrpcClient {
-    address: String,
-    client: ValidatorNodeClient<Channel>,
-    notification_retry_delay: Duration,
-    notification_retries: u32,
-}
-
-impl GrpcClient {
-    pub fn new(
-        network: ValidatorPublicNetworkConfig,
-        options: NodeOptions,
-    ) -> Result<Self, GrpcError> {
-        let address = network.http_address();
-        let channel = Channel::from_shared(address.clone())?
-            .connect_timeout(options.send_timeout)
-            .timeout(options.recv_timeout)
-            .connect_lazy();
-        let client = ValidatorNodeClient::new(channel)
-            .max_encoding_message_size(MAX_MESSAGE_SIZE)
-            .max_decoding_message_size(MAX_MESSAGE_SIZE);
-
-        Ok(Self {
-            address,
-            client,
-            notification_retry_delay: options.notification_retry_delay,
-            notification_retries: options.notification_retries,
-        })
-    }
-
-    /// Returns whether this gRPC status means the server stream should be reconnected to, or not.
-    /// Logs a warning on unexpected status codes.
-    fn is_retryable(status: &Status) -> bool {
-        match status.code() {
-            Code::DeadlineExceeded | Code::Aborted | Code::Unavailable | Code::Unknown => {
-                info!("Notification stream interrupted: {}; retrying", status);
-                true
-            }
-            Code::Ok
-            | Code::Cancelled
-            | Code::NotFound
-            | Code::AlreadyExists
-            | Code::ResourceExhausted => {
-                warn!("Unexpected gRPC status: {}; retrying", status);
-                true
-            }
-            Code::InvalidArgument
-            | Code::PermissionDenied
-            | Code::FailedPrecondition
-            | Code::OutOfRange
-            | Code::Unimplemented
-            | Code::Internal
-            | Code::DataLoss
-            | Code::Unauthenticated => {
-                warn!("Unexpected gRPC status: {}", status);
-                false
-            }
-        }
-    }
-}
-
-macro_rules! client_delegate {
-    ($self:ident, $handler:ident, $req:ident) => {{
-        debug!(request = ?$req, "sending gRPC request");
-        let request_inner = $req.try_into().map_err(|_| NodeError::GrpcError {
-            error: "could not convert request to proto".to_string(),
-        })?;
-        let request = Request::new(request_inner);
-        match $self
-            .client
-            .$handler(request)
-            .await
-            .map_err(|s| NodeError::GrpcError {
-                error: format!(
-                    "remote request [{}] failed with status: {:?}",
-                    stringify!($handler),
-                    s
-                ),
-            })?
-            .into_inner()
-            .inner
-            .ok_or(NodeError::GrpcError {
-                error: "missing body from response".to_string(),
-            })? {
-            Inner::ChainInfoResponse(response) => {
-                Ok(response.try_into().map_err(|err| NodeError::GrpcError {
-                    error: format!("failed to marshal response: {}", err),
-                })?)
-            }
-            Inner::Error(error) => {
-                Err(bincode::deserialize(&error).map_err(|err| NodeError::GrpcError {
-                    error: format!("failed to marshal error message: {}", err),
-                })?)
-            }
-        }
-    }};
-}
-
-#[async_trait]
-impl ValidatorNode for GrpcClient {
-    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
-    async fn handle_block_proposal(
-        &mut self,
-        proposal: data_types::BlockProposal,
-    ) -> Result<linera_core::data_types::ChainInfoResponse, NodeError> {
-        client_delegate!(self, handle_block_proposal, proposal)
-    }
-
-    #[instrument(target = "grpc_client", skip_all, fields(address = self.address))]
-    async fn handle_lite_certificate(
-        &mut self,
-        certificate: data_types::LiteCertificate<'_>,
-        delivery: CrossChainMessageDelivery,
-    ) -> Result<linera_core::data_types::ChainInfoResponse, NodeError> {
-        let wait_for_outgoing_messages = delivery.wait_for_outgoing_messages();
-        let request = HandleLiteCertificateRequest {
-            certificate,
-            wait_for_outgoing_messages,
-        };
-        client_delegate!(self, handle_lite_certificate, request)
-    }
-
-    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
-    async fn handle_certificate(
-        &mut self,
-        certificate: data_types::Certificate,
-        blobs: Vec<data_types::HashedValue>,
-        delivery: CrossChainMessageDelivery,
-    ) -> Result<linera_core::data_types::ChainInfoResponse, NodeError> {
-        let wait_for_outgoing_messages = delivery.wait_for_outgoing_messages();
-        let request = HandleCertificateRequest {
-            certificate,
-            blobs,
-            wait_for_outgoing_messages,
-        };
-        client_delegate!(self, handle_certificate, request)
-    }
-
-    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
-    async fn handle_chain_info_query(
-        &mut self,
-        query: linera_core::data_types::ChainInfoQuery,
-    ) -> Result<linera_core::data_types::ChainInfoResponse, NodeError> {
-        client_delegate!(self, handle_chain_info_query, query)
-    }
-
-    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
-    async fn subscribe(&mut self, chains: Vec<ChainId>) -> Result<NotificationStream, NodeError> {
-        let notification_retry_delay = self.notification_retry_delay;
-        let notification_retries = self.notification_retries;
-        let mut retry_count = 0;
-        let subscription_request = SubscriptionRequest {
-            chain_ids: chains.into_iter().map(|chain| chain.into()).collect(),
-        };
-        let mut client = self.client.clone();
-
-        // Make the first connection attempt before returning from this method.
-        let mut stream = Some(
-            client
-                .subscribe(subscription_request.clone())
-                .await
-                .map_err(|status| NodeError::SubscriptionFailed {
-                    status: status.to_string(),
-                })?
-                .into_inner(),
-        );
-
-        // A stream of `Result<grpc::Notification, tonic::Status>` that keeps calling
-        // `client.subscribe(request)` endlessly and without delay.
-        let endlessly_retrying_notification_stream = stream::unfold((), move |()| {
-            let mut client = client.clone();
-            let subscription_request = subscription_request.clone();
-            let mut stream = stream.take();
-            async move {
-                let stream = if let Some(stream) = stream.take() {
-                    future::Either::Right(stream)
-                } else {
-                    match client.subscribe(subscription_request.clone()).await {
-                        Err(err) => future::Either::Left(stream::iter(iter::once(Err(err)))),
-                        Ok(response) => future::Either::Right(response.into_inner()),
-                    }
-                };
-                Some((stream, ()))
-            }
-        })
-        .flatten();
-
-        // The stream of `Notification`s that inserts increasing delays after retriable errors, and
-        // terminates after unexpected or fatal errors.
-        let notification_stream = endlessly_retrying_notification_stream
-            .map(|result| {
-                Notification::try_from(result?).map_err(|err| {
-                    let message = format!("Could not deserialize notification: {}", err);
-                    tonic::Status::new(Code::Internal, message)
-                })
-            })
-            .take_while(move |result| {
-                let Err(status) = result else {
-                    retry_count = 0;
-                    return future::Either::Left(future::ready(true));
-                };
-                if !Self::is_retryable(status) || retry_count >= notification_retries {
-                    return future::Either::Left(future::ready(false));
-                }
-                let delay = notification_retry_delay.saturating_mul(retry_count);
-                retry_count += 1;
-                future::Either::Right(async move {
-                    tokio::time::sleep(delay).await;
-                    true
-                })
-            })
-            .filter_map(|result| future::ready(result.ok()));
-
-        Ok(Box::pin(notification_stream))
-    }
-
-    #[instrument(target = "grpc_client", skip_all, err, fields(address = self.address))]
-    async fn get_version_info(&mut self) -> Result<VersionInfo, NodeError> {
-        Ok(self.client.get_version_info(()).await?.into_inner().into())
-    }
-}
-
-#[async_trait]
-impl MassClient for GrpcClient {
-    #[instrument(skip_all, err)]
-    async fn send(
-        &mut self,
-        requests: Vec<RpcMessage>,
-        max_in_flight: usize,
-    ) -> Result<Vec<RpcMessage>, MassClientError> {
-        let client = &mut self.client;
-        let responses = stream::iter(requests)
-            .map(|request| {
-                let mut client = client.clone();
-                async move {
-                    let response = match request {
-                        RpcMessage::BlockProposal(proposal) => {
-                            let request = Request::new((*proposal).try_into()?);
-                            client.handle_block_proposal(request).await?
-                        }
-                        RpcMessage::Certificate(request) => {
-                            let request = Request::new((*request).try_into()?);
-                            client.handle_certificate(request).await?
-                        }
-                        msg => panic!("attempted to send msg: {:?}", msg),
-                    };
-                    match response
-                        .into_inner()
-                        .inner
-                        .ok_or(ProtoConversionError::MissingField)?
-                    {
-                        Inner::ChainInfoResponse(chain_info_response) => {
-                            Ok(Some(RpcMessage::ChainInfoResponse(Box::new(
-                                chain_info_response.try_into()?,
-                            ))))
-                        }
-                        Inner::Error(error) => {
-                            let error = bincode::deserialize::<NodeError>(&error)
-                                .map_err(ProtoConversionError::BincodeError)?;
-                            error!(?error, "received error response");
-                            Ok(None)
-                        }
-                    }
-                }
-            })
-            .buffer_unordered(max_in_flight)
-            .filter_map(
-                |result: Result<Option<_>, MassClientError>| async move { result.transpose() },
-            )
-            .collect::<Vec<_>>()
-            .await
-            .into_iter()
-            .collect::<Result<Vec<_>, _>>()?;
-        Ok(responses)
-    }
-}
-
 /// Types which are proxyable and expose the appropriate methods to be handled
 /// by the `GrpcProxy`
 pub trait Proxyable {
@@ -954,11 +632,11 @@ impl Proxyable for ChainInfoQuery {
 
 impl Proxyable for CrossChainRequest {
     fn chain_id(&self) -> Option<ChainId> {
-        use grpc::cross_chain_request::Inner;
+        use super::api::cross_chain_request::Inner;
 
         match self.inner.as_ref()? {
-            Inner::UpdateRecipient(grpc::UpdateRecipient { recipient, .. })
-            | Inner::ConfirmUpdatedRecipient(grpc::ConfirmUpdatedRecipient { recipient, .. }) => {
+            Inner::UpdateRecipient(api::UpdateRecipient { recipient, .. })
+            | Inner::ConfirmUpdatedRecipient(api::ConfirmUpdatedRecipient { recipient, .. }) => {
                 recipient.clone()?.try_into().ok()
             }
         }
