+++ linera-rpc/src/simple/mod.rs
@@ -1,12 +1,16 @@
 // Copyright (c) Zefchain Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+mod client;
 mod codec;
-mod network;
 mod node_provider;
+#[cfg(with_server)]
+mod server;
 mod transport;
 
+pub use client::*;
 pub use codec::*;
-pub use network::*;
 pub use node_provider::*;
+#[cfg(with_server)]
+pub use server::*;
 pub use transport::*;
